third_party/WebKit/LayoutTests/fast/dom/shadow/host-context-class-change-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Class change affecting a node in a distributed node's subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(b, null).backgroundColor is transparent
PASS getComputedStyle(b, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
